update dashboard register to use certificate when connect to dashboard move dashboard tests to dashboard_register_test
